better file upload status and modal

+++ composeApp/src/desktopMain/kotlin/dev/ploiu/file_server_ui_new/pages/FolderPage.kt
@@ -4,25 +4,14 @@ import androidx.compose.foundation.ContextMenuArea
 import androidx.compose.foundation.ContextMenuItem
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.TooltipArea
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Error
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Snackbar
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
+import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -174,8 +163,10 @@ fun FolderPage(
             bodyText = "Are you sure you want to delete this folder? Type the name to confirm",
             confirmText = "Delete",
             onConfirm = {
-                folderActionState = NoFolderAction()
-                view.deleteFolder(action.folder)
+                if (it == action.folder.name) {
+                    folderActionState = NoFolderAction()
+                    view.deleteFolder(action.folder)
+                }
             },
             onCancel = { folderActionState = NoFolderAction() },
         )
